// exampleClient/src/app/email-builder/email-editor/email-variable/property-type.ts
export enum PropertyType {
  TEXT = 'Text',
  MULTILINE = 'Multi-line Text',
  DATE = 'Date',
  NUMBER = 'Number',
  EMAIL = 'Email',
  PHONE = 'Phone',
  PERCENTAGE = 'Percentage',
  HYPERLINK = 'Hyperlink',
  CURRENCY = 'Currency',
  IMAGE = 'Image',
  LIST = 'List',
  CLICKABLE_IMAGE = 'Clickable Image',
  LIST_OF_IMAGES = 'List of Images',
}